+++ src/main/java/techcourse/myblog/config/MyblogConfiguration.java
@@ -1,11 +1,14 @@
-package techcourse.myblog.web.support;
+package techcourse.myblog.config;
 
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import techcourse.myblog.interceptor.LoginInterceptor;
+import techcourse.myblog.web.support.UserSessionArgumentResolver;
 
 import java.util.List;
 
@@ -29,5 +32,10 @@ public class MyblogConfiguration implements WebMvcConfigurer {
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
         resolvers.add(new UserSessionArgumentResolver());
+    }
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
     }
 }
