+++ src/main/resources/templates/common/dropdown_login.html
@@ -4,7 +4,7 @@
             <img alt="Brown" class="profile-img img-fluid"
                  src="https://avatars3.githubusercontent.com/u/46308949?s=460&v=4">
             <div class="user-info text-white">
-                <span class="name pdd-right-5" th:text="${session.userName}"></span>
+                <span class="name pdd-right-5" th:text="${session.author}"></span>
             </div>
         </a>
         <ul class="dropdown-menu">
